refactor : 코드 리뷰 반영 - Distance 객체 생성
